requirements.txt
allure-pytest==2.14.0
allure-python-commons==2.14.0
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
Faker==37.1.0
h11==0.14.0
idna==3.10
iniconfig==2.1.0
outcome==1.3.0.post0
packaging==25.0
pluggy==1.5.0
PySocks==1.7.1
pytest==8.3.5
python-dotenv==1.1.0
requests==2.32.3
selenium==4.31.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
